serach import

+++ models/ImportRecord.js
@@ -13,8 +13,8 @@
 
 global.ImportRecord = sequelize.define('improtRecord',{
   productName: {type: Sequelize.STRING},
-  salePrice:{type: Sequelize.INTEGER},
-  costPrice:{type: Sequelize.INTEGER},
+  salePrice:{type: Sequelize.FLOAT},
+  costPrice:{type: Sequelize.FLOAT},
   weight:{type: Sequelize.STRING},
   taste :{type: Sequelize.STRING},
   addStockNum:{type: Sequelize.INTEGER,defaultValue:0},
